requirements.txt
fastapi==0.110.0
python-multipart==0.0.9
azure-ai-formrecognizer==3.3.2
azure-core==1.30.1
python-dotenv==1.0.1
PyMuPDF==1.24.0
Pillow==10.2.0
uvicorn==0.28.0
gunicorn==21.2.0
